requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2020.11.8
cloudpickle==1.6.0
concurrent-iterator==0.2.6
cycler==0.10.0
dask==2020.12.0
dataclasses==0.8
decorator==4.4.2
future==0.18.2
html5lib==0.9999999
imageio==2.9.0
importlib-metadata==3.3.0
joblib==0.17.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==2.2.2
networkx==2.5
numpy==1.16.1
olefile==0.46
pandas==0.20.3
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.14.2
scikit-learn==0.19.1
scipy==0.19.1
six==1.15.0
sklearn==0.0
tensorboardX==2.1
tensorflow==1.2.1
tensorflow-gpu==1.2.1
threadpoolctl==2.1.0
tifffile==2020.9.3
toolz==0.11.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.54.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
